Changed energy consumption method

+++ app/src/main/java/com/example/edgedashanalytics/advanced/common/FrameResult.java
@@ -12,10 +12,10 @@ public class FrameResult {
     public long workerTime;
     public long turnaround;
     public long queueSize;
-    public long powerConsumed;
+    public int powerConsumed;
 
     public FrameResult(long timestamp, int frameNum, int cameraFrameNum, int workerNum, boolean isInner,
-                       long workerTime, long processTime, long networkTime, long turnaround, long queueSize, long powerConsumed) {
+                       long workerTime, long processTime, long networkTime, long turnaround, long queueSize, int powerConsumed) {
         //Log.v(TAG, "processTime = " + processTime);
         this.timestamp = timestamp;
         this.frameNum = frameNum;
